Attempting to fix build issue by moving GetTheme into components
