+++ galasa-ui/src/tests/routes/authTokens.test.ts
@@ -7,7 +7,7 @@
 import * as AuthTokenRoute from '@/app/auth/tokens/route';
 import { AuthenticationAPIApi } from '@/generated/galasaapi';
 import { NextRequest } from 'next/server';
-import { GetTheme } from '@/utils/GetTheme';
+import { GetTheme } from '@/components/GetTheme';
 
 const mockAuthenticationApi = AuthenticationAPIApi as jest.Mock;
